Add `ManagedJobsNamespaceSelector` to Kueue configuration
When using `labelPolicy` with `None` value, the `managedJobsWithoutQueueName`
is enabled, consequently any job that doesn't have the kueue label set
will still be managed by Kueue. To reduce the scope of affected jobs
this commits set the `ManagedJobsNamespaceSelector` to only select the namespaces
with label `kueue.openshift.io/managed=true`.