Clean up e2e targets
If we run undeploy-ocp after each test suite,
we would get kueue assets removed and if another
test suite attempts to run, it would fail. Also,
in CI there is one specific post step to get the
must-gather, so we don't need to run it on each test target.
This commit removed the undeploy-ocp and must-gather from
the test targets that run on CI.