Manually sync operator image in the Bundle
This commit manually syncs the operator image in the bundle
to make sure our periodic jobs use the latest operator code
avaialable. This fix is needed while we work on identifying
what is wrong with the nudges. It also removes an unecessary
file that is created when generating the bundle.