Add support to allow-nominal-concurrency-shares-update annotation
This commit adds support to an annotation to the `PriorityLevelConfiguration`
asset, which allows the user to modify the `nominalConcurrencyShares`
field to values bigger or equal than 2 and less or equal than 5 and to 0,
which rejects any traffic to the extension API. This is needed in case the user
experiences issues with the API and want to tune it. Also, note that the default
value is being modifed to 2, which means this would allow 32 requests to be
handled concurrently.